test: split endpoint resolver tests
